[PLAT-11299]: Backup deletion improvements
Summary:
Added the changes to have 3 retries before actually calling out a backup deletion failure.
In the case of GCS backup deletion, we will now create batches of size 100 only, which does increase the backup deletion time, but it is mentioned as the best practice in Google Docs: https://cloud.google.com/storage/docs/batch
```
You should not include more than 100 calls in a single batch request. If you need to make more calls than that, use multiple batch requests. The total batch request payload must be less than 10MB.
```

On side note: I was able to verify that if we try to delete non-null zero blobs list, we get this below error, and fixed it by adding a check to return from func if blob size is zero.
```
java.lang.IllegalStateException: null
        at com.google.common.base.Preconditions.checkState(Preconditions.java:496)
        at com.google.api.client.util.Preconditions.checkState(Preconditions.java:79)
        at com.google.api.client.googleapis.batch.BatchRequest.execute(BatchRequest.java:231)
        at com.google.cloud.storage.spi.v1.HttpStorageRpc$DefaultRpcBatch.submit(HttpStorageRpc.java:205)
        at com.google.cloud.storage.StorageBatch.submit(StorageBatch.java:149)
```

Test Plan:
Tested manually by
 - failing the backups deletion once to check that backup deletion failed state is updated only after 3 retries.
 - deleting empty backup and it passed.
 - deleting a backup of 10000 obect by creating more than 4000 tables.

Reviewers: vpatibandla, vkumar, kkg, #yba-api-review!

Reviewed By: vpatibandla

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D30425

+++ managed/src/main/java/com/yugabyte/yw/commissioner/BackupGarbageCollector.java
@@ -76,6 +76,7 @@ public class BackupGarbageCollector {
   private static final String GCS = Util.GCS;
   private static final String S3 = Util.S3;
   private static final String NFS = Util.NFS;
+  private static final int BACKUP_DELETION_MAX_RETRIES_COUNT = 3;
 
   public static final Gauge DELETE_BACKUP_FAILURE =
       Gauge.build("ybp_delete_backup_failure", "Count of failed delete backup attempt")
@@ -106,6 +107,7 @@ public class BackupGarbageCollector {
 
   public void start() {
     Duration gcInterval = this.gcRunInterval();
+    handleDeleteInProgressBackups();
     platformScheduler.schedule(
         getClass().getSimpleName(), Duration.ZERO, gcInterval, this::scheduleRunner);
   }
@@ -116,6 +118,18 @@ public class BackupGarbageCollector {
 
   private int getDeleteExpiredBackupMaxGCCount() {
     return confGetter.getGlobalConf(GlobalConfKeys.deleteExpiredBackupMaxGCSize);
+  }
+
+  public void handleDeleteInProgressBackups() {
+    for (Customer customer : Customer.getAll()) {
+      List<Backup> deleteInProgressBackups =
+          Backup.findAllBackupWithState(
+              customer.getUuid(), Collections.singletonList(BackupState.DeleteInProgress));
+      if (deleteInProgressBackups != null) {
+        deleteInProgressBackups.forEach(
+            backup -> backup.transitionState(BackupState.QueuedForDeletion));
+      }
+    }
   }
 
   void scheduleRunner() {
@@ -307,11 +321,27 @@ public class BackupGarbageCollector {
           case AZ:
             CloudUtil cloudUtil = storageUtilFactory.getCloudUtil(customerConfig.getName());
             backupLocations = BackupUtil.getBackupLocations(backup);
-            cloudUtil.deleteKeyIfExists(customerConfig.getDataObject(), backupLocations.get(0));
-            cloudUtil.deleteStorage(customerConfig.getDataObject(), backupLocations);
-            backup.delete();
-            deletedSuccessfully = true;
-            log.info("Backup {} is successfully deleted", backupUUID);
+            int numRetries = 0;
+            long sleepTimeInMilliSeconds = 5000;
+            while (numRetries < BACKUP_DELETION_MAX_RETRIES_COUNT && !deletedSuccessfully) {
+              if (cloudUtil.deleteKeyIfExists(
+                      customerConfig.getDataObject(), backupLocations.get(0))
+                  && cloudUtil.deleteStorage(customerConfig.getDataObject(), backupLocations)) {
+                deletedSuccessfully = true;
+              }
+              if (!deletedSuccessfully) {
+                Thread.sleep(sleepTimeInMilliSeconds);
+                sleepTimeInMilliSeconds = sleepTimeInMilliSeconds * 2;
+              }
+              numRetries++;
+            }
+            if (deletedSuccessfully) {
+              backup.delete();
+              log.info("Backup {} is successfully deleted", backupUUID);
+            } else {
+              backup.transitionState(Backup.BackupState.FailedToDelete);
+              log.info("Backup {} deletion failed", backupUUID);
+            }
             break;
           case NFS:
             if (isUniversePresent(backup)) {
